+++ R/notation.R
@@ -8,8 +8,8 @@
 #' These functions change a vector of strings between the specification notations.
 #' 
 #' * `switch_notation()` is the worker function.
-#' * `arrow_to_paren()` switches from arrow to parenthetical specification notations.
-#' * `paren_to_arrow()` switches from parenthetical to arrow specification notations.
+#' * `arrow_to_paren()` switches from arrow to parenthetical specification notation.
+#' * `paren_to_arrow()` switches from parenthetical to arrow specification notation.
 #'
 #' @return An object of same length as `strings` with switched notation.
 #' 
@@ -20,6 +20,7 @@
 #' @param new_end the end of specification notation in output
 #'
 #' @examples
+#' 
 #' @name switch-notation
 NULL
 
@@ -68,3 +69,69 @@ paren_to_arrow <- function(x,
                   old_start = old_start, old_end = old_end,
                   new_start = new_start, new_end = new_end)
 }
+
+
+#' Change specification notation on matrix row or column names
+#' 
+#' Switch notation style 
+#' in row or column names of matrices or a list of matrices.
+#' Notation style is either arrow ("a -> b") or parenthetical ("b \[a\]")
+#' 
+#' The suffix "_byname" indicates that this function behaves like functions in the `byname` package.
+#' Specifically, it can work with a single matrix or a list of matrices supplied to the `m` argument.
+#' 
+#' * `arrow_to_paren_byname()` switches from arrow to parenthetical specification notation for a matrix or list of matrices.
+#' * `paren_to_arrow_byname()` switches from parenthetical to arrow specification notation for a matrix or list of matrices.
+#'
+#' @param m a single matrix or a list of matrices.
+#' @param margin the margin over which the notation switch should be made:
+#'               `1` for rows, `2` for columns, or `c(1, 2)` (the default) for both rows and columns.
+#'
+#' @return a version of `m` with switched notation
+#'
+#' @seealso `arrow_to_paren()`, `paren_to_arrow()`, and `switch_notation()`
+#' 
+#' @examples
+#' 
+#' @name switch-notation-byname
+NULL
+
+#' @export
+#' @rdname switch-notation-byname
+arrow_to_paren_byname <- function(m, margin = c(1, 2)) {
+  switch_notation_byname(m, margin, arrow_to_paren)
+}
+
+
+
+#' @export
+#' @rdname switch-notation-byname
+paren_to_arrow_byname <- function(m, margin = c(1, 2)) {
+  switch_notation_byname(m, margin, paren_to_arrow)
+}
+
+
+switch_notation_byname <- function(m, margin, func) {
+  assertthat::assert_that(all(margin %in% c(1, 2)), msg = "margin must be 1, 2, or both.")
+  
+  out <- m
+  if (2 %in% margin) {
+    # Transpose the matrices
+    transposed <- matsbyname::transpose_byname(out)
+    # re-call with margin = 1 to change from arrow to paren notation on the rows (which are really columns)
+    switched <- switch_notation_byname(transposed, margin = 1, func)
+    # Transpose
+    out <- matsbyname::transpose(switched)
+  }
+  if (1 %in% margin) {
+    # Get the row names
+    old_rownames <- matsbyname::getrownames_byname(out)
+    # call func all all row names to create new row names
+    new_rownames <- func(old_rownames)
+    # Set row names to the new row names
+    out <- matsbyname::setrownames_byname(out, new_rownames)
+  }
+  # Return the result
+  return(out)
+}
+